requirements.txt
astroid==1.3.8
blaze==0.8.3
boto3==1.1.4
botocore==1.2.5
datashape==0.4.7
decorator==4.0.4
docutils==0.12
futures==2.2.0
jmespath==0.8.0
linecache2==1.0.0
logilab-common==1.0.2
multipledispatch==0.4.8
networkx==1.10
numpy==1.10.1
odo==0.3.4
ordereddict==1.1
pandas==0.17.1
psutil==3.3.0
psycopg2==2.6.1
py==1.4.30
Pygments==2.0.2
pylint==1.4.4
pytest==2.8.0
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
SQLAlchemy==1.0.9
toolz==0.7.4
traceback2==1.4.0
unittest2==1.1.0
virtualenv==13.1.2
wheel==0.26.0
wmctrl==0.2
python-dotenv==0.1.5
requests==2.10.0
redshift-sqlalchemy==0.4.1
